refactoring backend imports after merge with sprint 2
