+++ backend/JavaWebService/src/main/java/com/kainos/ea/WebServiceApplication.java
@@ -2,6 +2,8 @@ package com.kainos.ea;
 import com.kainos.ea.db.CapabilityMapper;
 import com.kainos.ea.db.RoleMapper;
 import com.kainos.ea.db.RolesMapper;
+import com.kainos.ea.resources.CapabilitiesController;
+import com.kainos.ea.resources.CapabilitiesService;
 import com.kainos.ea.services.BandService;
 import com.kainos.ea.services.RoleService;
 import com.kainos.ea.controllers.*;
