+++ backend/JavaWebService/src/main/java/com/kainos/ea/models/Capability.java
@@ -1,4 +1,4 @@
-package com.kainos.ea.resources;
+package com.kainos.ea.models;
 
 public class Capability {
     private int capabilityID;
