+++ backend/JavaWebService/src/main/java/com/kainos/ea/resources/CapabilitiesController.java
@@ -1,5 +1,8 @@
 package com.kainos.ea.resources;
 
+import com.kainos.ea.models.Capability;
+import com.kainos.ea.models.Role;
+
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
